edge/services/vitals/requirements.txt
paho-mqtt>=2.0.0
rpi-lgpio
grove.py
seeed-python-dht